chore: rename PlatformPayView
