+++ android/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
 <manifest
-    package="com.adyenreactnativesdk"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application>
