--- android/src/main/AndroidManifestNew.xml
@@ -1,8 +1,0 @@
-<manifest
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <service android:name="com.adyenreactnativesdk.component.dropin.AdvancedCheckoutService" android:exported="false" />
-        <service android:name="com.adyenreactnativesdk.component.dropin.SessionCheckoutService" android:exported="false" />
-    </application>
-</manifest>
