+++ src/index.ts
@@ -1,6 +1,3 @@
-export { default as PlatformPayView } from './specs/PlatformPayViewNativeComponent';
-export * from './specs/PlatformPayViewNativeComponent';
-
 export * from './components/GooglePayButton';
 export * from './components/ApplePayButton';
