+++ src/specs/NativePlatformPayView.ts
