+++ tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig",
-  "exclude": ["example", "lib", "**/__mocks__/*", "**/__tests__/*"]
+  "exclude": ["example", "plugin", "lib", "**/__mocks__/*", "**/__tests__/*"]
 }
